backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
packaging==21.3
pyparsing==3.0.9
pytz==2022.2.1
redis==4.3.4
sqlparse==0.4.2
wrapt==1.14.1
